Moved tinyxml2 and imgui to Dependencies/
