+++ Sources/Overload/OvUI/premake5.lua
@@ -3,7 +3,7 @@ project "OvUI"
 	language "C++"
 	cppdialect "C++20"
 	files { "**.h", "**.inl", "**.cpp", "**.lua" }
-	includedirs { "include", dependdir .. "glfw/include", dependdir .. "glew/include", "%{wks.location}/OvMaths/include", "%{wks.location}/OvTools/include", "%{wks.location}/../ImGui/" }
+	includedirs { "include", dependdir .. "glfw/include", dependdir .. "glew/include", dependdir .. "ImGui", "%{wks.location}/OvMaths/include", "%{wks.location}/OvTools/include", "%{wks.location}/../ImGui/" }
 	targetdir (outputdir .. "%{cfg.buildcfg}/%{prj.name}")
 	objdir (objoutdir .. "%{cfg.buildcfg}/%{prj.name}")
 	characterset ("MBCS")
